+++ 2019/asteroid/slope.go
@@ -5,13 +5,13 @@ import (
 )
 
 type slope struct {
-	dx int32
-	dy int32
+	dx int64
+	dy int64
 }
 
 func calculateSlope(a math.Point, b math.Point) slope {
-	dx := a.X - b.X
-	dy := a.Y - b.Y
+	dx := int64(a.X - b.X)
+	dy := int64(a.Y - b.Y)
 	gcd := math.GCD(math.Abs(dx), math.Abs(dy))
 	return slope{
 		dx / gcd,
